Replaced auth initialization with custom hook in App.jsx, and organized the imports
