+++ doggii-frontend/src/App.jsx
@@ -1,5 +1,11 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+
+import { useAuth } from './hooks/useAuth';
+
 import Login from './pages/Login';
+import OptionsForms from './pages/OptionsForms';
+import FormAdopter from './pages/FormAdopter';
+import FormShelter from './pages/FormShelter';
 import PlatformLayout from './layouts/PlatformLayout';
 import Dashboard from './pages/Dashboard/Dashboard';
 import Profile from './pages/Profile/Profile';
@@ -7,87 +13,20 @@ import Pets from './pages/Pets/Pets';
 import Applications from './pages/Applications/Applications';
 import ApplicationDetail from './pages/Applications/ApplicationViews/AplicationDetail.jsx';
 import AdoptionForm from './pages/AdoptionForm';
-import OptionsForms from './pages/OptionsForms';
-import FormAdopter from './pages/FormAdopter';
-import FormShelter from './pages/FormShelter';
 import Users from './pages/Users/Users';
 import Reports from './pages/Reports/Reports';
-import LandingLayout from './layouts/LandingLayout';
 import PetPost from './pages/Pets/PetPost/PetPost';
-import PetEdit from './pages/Pets/PetEdit/PetEdit';
 import PetDetails from './pages/Pets/PetDetails/PetDetails';
-import { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import axios from 'axios';
-import { jwtDecode } from "jwt-decode";
-import { setUserInfo } from './redux/slices/userSlice';
-import { loginSuccess } from './redux/slices/authSlice';
+import PetEdit from './pages/Pets/PetEdit/PetEdit';
+import LandingLayout from './layouts/LandingLayout';
+
 
 function App() {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    const initializeAuth = async () => {
-      const token = localStorage.getItem('accessToken');
-      
-      if (!token) return;
-
-      try {
-        // Verificar si el token ha expirado
-        const decodedToken = jwtDecode(token);
-        const currentTime = Date.now() / 1000;
-        if (decodedToken.exp < currentTime) {
-          localStorage.removeItem('accessToken');
-          delete axios.defaults.headers.common['Authorization'];
-          return;
-        }
-
-        // Configurar axios con el token válido
-        axios.defaults.headers.common['Authorization'] = `Bearer ${token}`;
-        
-        // Obtener el rol del token y llamar al endpoint correspondiente
-        const userRole = decodedToken.role.toLowerCase();
-        let endpoint;
-        
-        switch(userRole) {
-          case 'adopter':
-            endpoint = '/api/adopter/' + decodedToken.id;
-            break;
-          case 'shelter':
-            endpoint = '/api/shelter/' + decodedToken.id;
-            break;
-          case 'admin':
-            endpoint = '/api/admin/' + decodedToken.id;
-            break;
-          default:
-            throw new Error('Rol de usuario no válido');
-        }
-        
-        const response = await axios.get(import.meta.env.VITE_BACKEND_URI + endpoint);
-        
-        console.log('Respuesta del backend:', response.data);
-        
-        if (response.data.success) {
-          dispatch(setUserInfo({
-            ...response.data.payload,
-            _id: decodedToken.id
-          }));
-          dispatch(loginSuccess(token));
-        }
-      } catch (error) {
-        console.error('Error al recuperar la sesión:', error);
-        localStorage.removeItem('accessToken');
-        delete axios.defaults.headers.common['Authorization'];
-      }
-    };
-
-    initializeAuth();
-  }, [dispatch]);
+  useAuth();
 
   return (
     <BrowserRouter>
       <Routes>
-        {/* <Route path='/' element={<Home />} /> */}
         <Route path='/login' element={<Login />} />
         <Route path='/options-forms' element={<OptionsForms />} />
         <Route path='/form-adopter' element={<FormAdopter />} />
